chore: update .csproj version properties, add XML docs, alias LuceneDirectory, and use ConfigureAwait(false) in ReplicationClientService
