--- src/Lucene.Net.Extensions.ReplicationClient/Lucene.Net.Extensions.ReplicationClient.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <!-- Core Lucene.NET Replicator package -->
-    <PackageReference Include="Lucene.Net" Version="$(LuceneNetVersion)" />
-    <PackageReference Include="Lucene.Net.Replicator" Version="$(LuceneNetVersion)" />
-
-    <!-- Microsoft.Extensions packages -->
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="$(MicrosoftExtensionsVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="$(MicrosoftExtensionsVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="$(MicrosoftExtensionsVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="$(MicrosoftExtensionsVersion)" />
-  </ItemGroup>
-
-</Project>
